pyproject.toml
[tool.poetry]
name = "hackernews-fill-gaps"
version = "0.1.0"
description = ""
authors = ["Thomas Dziedzic <thomasdziedzic0@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
snowflake-connector-python = "^2.3.8"
python-dotenv = "^0.15.0"
bpython = "^0.20.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
